refactor: merged data_processor with data_collector for locality of behavior
