+++ modules/updater/data_processing/data_collector.py
@@ -1,11 +1,16 @@
 import pandas as pd
 import streamlit as st
 
-from modules.updater.data_processing.data_processor import (
-    build_dataframe,
-    process_records,
-)
+from modules.updater.data_processing.site_files import set_filename_from_link
+from modules.updater.sites.JobSite import JobSite
+from modules.updater.sites.SiteFactory import SiteFactory
 from modules.websites import search_links
+
+
+def build_dataframe(records):
+    """Convert list of job records to pandas DataFrame"""
+    records_matrix = [item.to_dict() for sublist in records for item in sublist]
+    return pd.DataFrame(records_matrix)
 
 
 def html_dataframe() -> pd.DataFrame:
@@ -32,3 +37,11 @@ def search_site(link: str) -> list:
         if "st" in globals():
             st.toast(f"**Run updater to process link:**\n{link}", icon="⚠️")
     return job_records
+
+
+def process_records(link: str):
+    """Process HTML soup into JobRecord objects"""
+    website: JobSite = SiteFactory.identify_website(link)
+    file_name = set_filename_from_link(link, website.file_extension)
+    records = website.records_list(html=file_name, link=link)
+    return [SiteFactory.single_record(website=website, record=record) for record in records] if records else []
